brain/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
httpx[socks]
langchain
langchain-google-genai
google-generativeai
chromadb
sentence-transformers
pypdf